Refactor to create `text.js`
